ReactComponentViewer/src/pyproject.toml
[tool.poetry]
name = "reactcomponentviewer"
version = "0.1.0"
description = ""
authors = ["ni-eminen <matiasnie2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
pywebview = "^3.6.2"
invoke = "^1.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
